+++ tests/Integration/PlasterTemplates/Sampler/DscCommunityPlasterTemplate.Integration.Tests.ps1
@@ -76,10 +76,6 @@ Describe 'DSC Community Plaster Template' {
             )
         }
 
-        AfterAll {
-            Remove-Item -Path $TestDrive -Recurse -Force
-        }
-
         It 'Should create a new module without throwing' {
             $invokePlasterParameters = @{
                 TemplatePath         = Join-Path -Path $importedModule.ModuleBase -ChildPath 'Templates/Sampler'