+++ tests/Integration/PlasterTemplates/Sampler/SimpleModuleNoBuildPlasterTemplate.Integration.Tests.ps1
@@ -46,10 +46,6 @@ Describe 'Simple Module No Build Plaster Template' {
             )
         }
 
-        AfterAll {
-            Remove-Item -Path $TestDrive -Recurse -Force
-        }
-
         It 'Should create a new module without throwing' {
             $invokePlasterParameters = @{
                 TemplatePath      = Join-Path -Path $importedModule.ModuleBase -ChildPath 'Templates/Sampler'